change sqlite to postgresql

+++ migrations/000002_messages.up.sql
@@ -1,9 +1,9 @@
 -- Messages table
 CREATE TABLE IF NOT EXISTS messages (
-    id INTEGER PRIMARY KEY,
-    chat_id INTEGER NOT NULL,
+    id SERIAL PRIMARY KEY,
+    chat_id BIGINT NOT NULL,
     text TEXT NOT NULL,
-    user_id INTEGER NOT NULL,
+    user_id BIGINT NOT NULL,
     username TEXT NOT NULL,
     timestamp TIMESTAMP NOT NULL,
     type TEXT NOT NULL DEFAULT 'text',
